0.0.69 - Texrue' Drag&Drop  into selected gameObject

+++ CollyraEngine/M_Scene.cpp
@@ -87,7 +87,7 @@ updateStatus M_Scene::Draw(float dt)
 			{
 				if (material != nullptr)
 				{
-					mesh->Render(transform->GetTGlobalTransform(), material->GetTexture(), material->GetColor());
+					mesh->Render(transform->GetTGlobalTransform(), (int)material->GetTexture(), material->GetColor());
 				}
 				else
 				{
